+++ ext/GoogleGenAIPromptingToolsExt.jl
@@ -1,0 +1,15 @@
+module GoogleGenAIPromptingToolsExt
+
+using GoogleGenAI
+using PromptingTools
+const PT = PromptingTools
+
+# TODO:
+function GoogleGenAI.generate_content(prompt_schema, api_key,
+        model_id,
+        conv_rendered;
+        http_kwargs,
+        api_kwargs...)
+end
+
+end # end of module